+++ src/main/java/com/charmeetchic/CharmeetChic/model/Notificaciones.java
@@ -6,21 +6,26 @@ import jakarta.persistence.*;
 public class Notificaciones {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
+
     private String destinatario;
     private String mensaje;
-    private String tipo; // ejemplo: EMAIL, SMS, PUSH
+    private String tipo; // EMAIL, SMS, PUSH, etc.
     private boolean enviado;
 
     // Getters y Setters
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
+
     public String getDestinatario() { return destinatario; }
     public void setDestinatario(String destinatario) { this.destinatario = destinatario; }
+
     public String getMensaje() { return mensaje; }
     public void setMensaje(String mensaje) { this.mensaje = mensaje; }
+
     public String getTipo() { return tipo; }
     public void setTipo(String tipo) { this.tipo = tipo; }
+
     public boolean isEnviado() { return enviado; }
     public void setEnviado(boolean enviado) { this.enviado = enviado; }
 }
